+++ components/dash-core-components/src/fragments/Dropdown.react.js
@@ -29,6 +29,7 @@ const RDProps = [
     'placeholder',
     'disabled',
     'optionHeight',
+    'maxHeight',
     'style',
     'className',
 ];
@@ -37,6 +38,7 @@ const Dropdown = props => {
     const {
         id,
         clearable,
+        searchable,
         multi,
         options,
         setProps,
@@ -86,6 +88,7 @@ const Dropdown = props => {
 
     useEffect(() => {
         if (
+            !searchable &&
             !isNil(sanitizedOptions) &&
             optionsCheck !== sanitizedOptions &&
             !isNil(value)