+++ dash/long_callback/__init__.py
@@ -1,2 +1,8 @@
-from .managers.celery_manager import CeleryLongCallbackManager  # noqa: F401,E402
-from .managers.diskcache_manager import DiskcacheLongCallbackManager  # noqa: F401,E402
+from .managers.celery_manager import (  # noqa: F401,E402
+    CeleryLongCallbackManager,
+    CeleryManager,
+)
+from .managers.diskcache_manager import (  # noqa: F401,E402
+    DiskcacheLongCallbackManager,
+    DiskcacheManager,
+)